+++ src/components/Loading/index.css
@@ -1,0 +1,6 @@
+.loading-wrapper {
+    /*background: rgba(0, 0, 0, 0.5);*/
+    background: red;
+    width: 100;
+    height: 100;
+}
