+++ src/components/Loading/index.tsx
@@ -1,0 +1,8 @@
+import "./index.css"
+
+const Loading = () => {
+    return <>
+    </>
+}
+
+export default Loading;
