+++ src/components/index.ts
@@ -11,10 +11,12 @@ import Sun from './Sun';
 import Skybox from './Skybox';
 import CameraController from './CameraController';
 import Model from './Model';
+import Loading from "./Loading";
 
 export {
     Scene, VrToggler, CelestialBody,
     SmallBodies, SideMenu, FiltersMenu,
     BodiesTable, TimeControls, Sun, Model,
-    ConstantSizeSphere, Skybox, CameraController
+    ConstantSizeSphere, Skybox, CameraController,
+    Loading
 };
